[FEAT/#153] 알람목록뷰 UseCase 수정
