+++ domain/src/main/kotlin/co/orange/domain/usecase/alarm/GetAlarmListUseCase.kt
@@ -3,12 +3,12 @@ package co.orange.domain.usecase.alarm
 import co.orange.domain.repository.HomeRepository
 import javax.inject.Inject
 
-class PatchAlarmToReadUseCase @Inject constructor(
+class GetAlarmListUseCase @Inject constructor(
     private val homeRepository: HomeRepository
 ) {
-    suspend operator fun invoke(alarmId: Long) =
+    suspend operator fun invoke() =
         runCatching {
-            val response = homeRepository.patchToReadAlarm(alarmId)
+            val response = homeRepository.getAlarmList()
             return@runCatching response
         }
 }
